requirements.txt
google-adk
psutil
litellm
google-generativeai
python-dotenv
deprecated
google-api-python-client
absl-py
